requirements.txt
python-telegram-bot==21.3
groq==0.9.0
python-dotenv==1.0.1
aiosqlite==0.20.0